refactor: 新規入会者向けメッセージを `new-members` ディレクトリに移動
