+++ messenger/src/messages/index.ts
@@ -1,9 +1,10 @@
 import { PullRequestOpenedEvent } from '@octokit/webhooks-types';
 
-import message01 from './01';
-import message02 from './02';
-import message03 from './03';
+import message01forNew from './new-members/01';
+import message02forNew from './new-members/02';
+import message03forNew from './new-members/03';
 
-const generateMessages = (webhook: PullRequestOpenedEvent) => [message01, message02, message03].map((f) => f(webhook));
+const generateMessagesForNew = (webhook: PullRequestOpenedEvent) =>
+	[message01forNew, message02forNew, message03forNew].map((f) => f(webhook));
 
-export default generateMessages;
+export { generateMessagesForNew };
